Team-Project-FinalBackend/Loyalty/src/main/java/org/loyalty/module/loyalty_module/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
web3>=6.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
python-multipart>=0.0.5 